BUG/MEDIUM: http: don't dump debug headers on MSG_ERROR
When the HTTP parser is in state HTTP_MSG_ERROR, we don't know if it was
already initialized or not. If the error happens before HTTP_MSG_RQBEFORE,
random offsets might be present and we don't want to display such random
strings in debug mode.

While it's theorically possible to randomly crash the process when running
in debug mode here, this bug was not tagged MAJOR because it would not
make sense to run in debug mode in production.

This fix must be backported to 1.5 and 1.4.
(cherry picked from commit 7d59e90473bbf3995ef11d146a7a4cf5c6d0c4b9)